Added all the recommendations, some tests left
