+++ mountpoint-s3-client/src/lib.rs
@@ -1,5 +1,6 @@
 mod build_info;
 pub mod checksums;
+mod endpoint_config;
 pub mod failure_client;
 mod imds_crt_client;
 pub mod mock_client;
@@ -7,12 +8,11 @@ mod object_client;
 mod s3_crt_client;
 mod util;
 
+pub use endpoint_config::{AddressingStyle, EndpointConfig};
 pub use imds_crt_client::ImdsCrtClient;
 pub use object_client::*;
 pub use s3_crt_client::head_bucket::HeadBucketError;
-pub use s3_crt_client::{
-    AddressingStyle, EndpointConfig, S3ClientAuthConfig, S3ClientConfig, S3CrtClient, S3RequestError,
-};
+pub use s3_crt_client::{S3ClientAuthConfig, S3ClientConfig, S3CrtClient, S3RequestError};
 
 #[cfg(test)]
 mod tests {
